Project re-organized in differents subfolders to separate tests from main program. Requires changes to the new rotues

+++ lib/file_handler_output.rb
@@ -2,10 +2,12 @@ class FileHandlerOutput
   OUTPUT_ROUTE = './out/out.txt'
 
   def initialize
-    unless File.exists?(OUTPUT_ROUTE)
-      Dir.mkdir('out')
-      File.write('./out/out.txt', '')
+    if File.exists?(OUTPUT_ROUTE)
+      File.delete('./out/out.txt')
+      Dir.delete('out')
     end
+    Dir.mkdir('out') unless Dir.exists?('out')
+    File.write('./out/out.txt', '') unless File.exists?('./out/out.txt')
   end
 
   def write_file(users_cliques)
